# swagger/swagger.yaml
swagger: '2.0'
info:
  version: "0.0.1"
  title: Thermostat Web API
  license:
    name: GPL v2 with Classpath Exception
    url: http://www.gnu.org/licenses
consumes: ['application/json']
produces: ['application/json', 'text/html; charset=utf-8']
paths:
...
  /agents/{agentId}:
    parameters:
      - $ref: '#/parameters/agent-id'    
    get:
      description: Gets information for an agent. Use 'all' as the agentId
      responses:
        401:
          $ref: '#/responses/unauthorized'
        200:
          description: Successful response
          schema:
              type: array
              items:
                $ref: '#/definitions/agent-info'     
# host-info
...
  /agents/{agentId}/host/cpu:
    parameters:
      - $ref: '#/parameters/agent-id'
    get:
      description: Gets cpu information for the host
      responses:
        401:
          $ref: '#/responses/unauthorized'      
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/host-cpu-info'
# memory-stats

# vm-info
...
definitions:
...
  agent-info:
    type: object
...
  host-cpu-info:
    type: object
...
parameters:
  agent-id:
    name: agentId
    in: path
    description: The agent identification number
    type: number
    required: true
...
responses:
  unauthorized:
    description: Access prohibited.